refactor: move RibbonIconKeys from ViewModels/Ribbon to Helpers
It's a plain constants class, not a ViewModel, and Assets/Styles in this project
are pure-XAML resource folders with no C# precedent. Helpers/ already holds this
exact shape of thing (e.g. TextFormatHelper), so it belongs there instead.
