+++ Mil.Paperwork.UI/ViewModels/Tabs/BaseReportTabViewModel.cs
@@ -3,6 +3,7 @@ using Mil.Paperwork.DataAccess.Services;
 using Mil.Paperwork.Domain.DataModels.ReportData;
 using Mil.Paperwork.Infrastructure.Enums;
 using Mil.Paperwork.Infrastructure.Services;
+using Mil.Paperwork.UI.Helpers;
 using Mil.Paperwork.UI.ViewModels.Ribbon;
 using System;
 using System.Collections.Generic;
